modif UtilisateurController if null + PUT {id}

+++ src/main/java/co/simplon/filrouge/controller/UtilisateurController.java
@@ -1,5 +1,7 @@
 package co.simplon.filrouge.controller;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -31,24 +33,54 @@ public class UtilisateurController {
 	}
 
 	@GetMapping(path = "/utilisateur/{id}")
-	public @ResponseBody Utilisateur getUtilisateur(@PathVariable Long id) throws Exception {
-		return utilisateurService.getUtilisateur(id);
+	public ResponseEntity<Utilisateur> getUtilisateur(@PathVariable Long id) throws Exception {
+		Utilisateur utilisateur = utilisateurService.getUtilisateur(id);
+		if (utilisateur == null) {
+		 return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+		}
+		else {
+		return ResponseEntity.status(HttpStatus.OK).body(utilisateur);
+		}
+		
 	}
 
 	@DeleteMapping(path = "/utilisateur/{id}")
-	public @ResponseBody void deleteUtilisateur(@PathVariable Long id) {
-		utilisateurService.delete(id);
+	public ResponseEntity<Utilisateur> deleteUtilisateur(@PathVariable Long id) throws Exception {
+		Utilisateur utilisateur = utilisateurService.getUtilisateur(id);
+		if (utilisateur == null) {
+			 return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+			}
+		else {
+			utilisateurService.delete(id);
+			return ResponseEntity.status(HttpStatus.OK).build();
+			}
 	}
 
 	@PostMapping(path = "/utilisateurs")
-	public ResponseEntity<?> createUtilisateur(@RequestBody Utilisateur user) throws Exception {
+	public ResponseEntity<?> createUtilisateur(@Valid @RequestBody Utilisateur user) throws Exception {
 		Utilisateur newUser = utilisateurService.addUtilisateur(user);
 		return ResponseEntity.status(HttpStatus.CREATED).body(newUser);
 	}
 
-	@PutMapping(path = "/utilisateurs")
-	public ResponseEntity<?> updateUtilisateur(@RequestBody Utilisateur user) throws Exception {
-		Utilisateur updatedUser = utilisateurService.editUtilisateur(user);
+	@PutMapping(path = "/utilisateur/{id}")
+	public ResponseEntity<Utilisateur> updateUtilisateur(@PathVariable Long id,@Valid @RequestBody Utilisateur user) throws Exception {
+		Utilisateur utilisateur = utilisateurService.getUtilisateur(id);
+		if (utilisateur == null) {
+			 return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+			}
+	
+		if(user.getNom()!=null)
+			utilisateur.setNom(user.getNom());
+		if(user.getPrenom()!=null)
+			utilisateur.setPrenom(user.getPrenom());
+		if(user.getMatricule()!=null);		
+			utilisateur.setMatricule(user.getMatricule());
+		if(user.getPassword()!=null)
+			utilisateur.setPassword(user.getPassword());
+		if(user.getHabilitation()!=null)
+			utilisateur.setHabilitation(user.getHabilitation());
+		
+		Utilisateur updatedUser = utilisateurService.editUtilisateur(utilisateur);
 		return ResponseEntity.status(HttpStatus.ACCEPTED).body(updatedUser);
 	}
 }
